xwax: Implement LUT loading/storing feature
Since the current LUT implementation is designed for Serato-style timecodes
with short LFSRs and not suitable for the Traktor MK2s with its 110-bit LFSR
and up to four million states, the computation of the hash map is taking
quite long (~10 seconds for the CD).

This is a dirty workaround of improving the startup time by not having
to wait for the LUT build every time. I am not happy with this solution,
but it does make the feature usable for now.

In the long run the we should look into pre-computing a Minimum Perfect
Hash Table, which should reduce the size of the LUT drastically.