xwax: Add a quadrature tracker
The current pitch filter works under the assumption that zero crossings
are never skipped. Unfortunately this happens when frequencies are very
high during e.g. scratching or backspins. Especially during backspins
the pitch filter was being starved by always feeding a quarter rotation
displacement, when in reality the displacement was a half rotation or
three quarters even.

This commit fixes this issue by mapping the quadrants to the complex
plane. The complex number of the last quadrant we were in, is compared
to the current complex number and the rotation angle is returned. This
rotation angle is then translated to timecode revolutions.

This fixes backspins and vastly improves sticker drift.